requirements.txt
# Embedding
sentence-transformers==3.0.1
faiss-cpu==1.8.0

# Backend
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
openai==1.51.0

# Utilities
numpy==1.26.4
